test: bits: fix printf formatting warning
We cast to the biggest signed integer size so it works in all used cases
(uint64_t is not used)

+++ test/src/misc/bits.c
@@ -26,7 +26,7 @@
     ssize_t a = (foo); \
     ssize_t b = (bar); \
     if(a != b) { \
-        printf("Failed (%s) line %d : %ld != %ld\n", psz_tag, __LINE__, a, b ); \
+        printf("Failed (%s) line %d : %" PRId64 " != %" PRId64 "\n", psz_tag, __LINE__, (int64_t)(a), (int64_t)(b) ); \
         return 1; \
     } } while( 0 )
